api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
torch
torchvision
pillow
numpy
python-dotenv
loguru
python-multipart